funcionamienot de prod./factura prod.

+++ api/post/producto.php
@@ -20,7 +20,7 @@ $data = json_decode(file_get_contents("php://input"));
 
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-    $validador=true;
+    $validador = true;
     $post = new Controller_Producto($GLOBALS['db']);
     $po = new Controller_bodega_has_producto($GLOBALS['db']);
     $p = new controller_bodega($GLOBALS['db']);
@@ -35,74 +35,106 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $fecha = date('Y-m-d');
     //$fecha = date('d-m-Y');
     //$validador=true;
-    
-    if ($post->validador_nombre($post->nombre_producto)!=null) {
-        $validador=false;
+
+    if ($post->validador_nombre($post->nombre_producto) != null) {
+        $validador = false;
         echo json_encode(
             array('message' => $post->validador_nombre($post->nombre_producto))
         );
     }
-    if ($post->Validador_valor_producto($post->valor_producto)!=null) {
-        $validador=false;
+    if ($post->Validador_valor_producto($post->valor_producto) != null) {
+        $validador = false;
         echo json_encode(
             array('message' => $post->Validador_valor_producto($post->valor_producto))
         );
     }
-    if ($po->Validador_cantidad_total($cantidad_total)!=null) {
-        $validador=false;
+    if ($po->Validador_cantidad_total($cantidad_total) != null) {
+        $validador = false;
         echo json_encode(
             array('message' => $po->Validador_cantidad_total($cantidad_total))
         );
     }
     if (empty($p->id_bodega)) {
-        $validador=false;
+        $validador = false;
         echo json_encode(
             array('message' => "ingrese una bodega")
         );
-    }else {
-        if ($p->buscar_id_bodega($p->id_bodega)!=false) {
-            $validador=false;
+    } else {
+        if ($p->buscar_id_bodega($p->id_bodega) != false) {
+            $validador = false;
             echo json_encode(
                 array('message' => "No existe la bodega")
             );
         }
     }
 
-    
-    if ($validador==true) {
+
+    if ($validador == true) {
         if ($post->create_producto()) {
 
-           
-            $id_producto=$post->obtener_el_ultimo_id();//manda
-            if ($po->create_bodega_has_producto($p->id_bodega,$id_producto,$cantidad_total)==false) {//funciona
+
+            $id_producto = $post->obtener_el_ultimo_id(); //manda
+            if ($po->create_bodega_has_producto($p->id_bodega, $id_producto, $cantidad_total) == false) { //funciona
                 echo json_encode(
                     array('message' => 'Error en ingreso de datos teniendo en cuenta el codigo del producto y el codigo de la bodega')
                 );
-            }else {
-                
-                //buscar el 
+            } else {
 
-                if ($di->create_detalle_inventario($post->nombre_producto,$cantidad_total,$post->valor_producto,1,$p->id_bodega,$id_producto,$fecha)==false) {//listo
+                //buscardor
+                $nombre = $di->buscardor_igual_producto($post->nombre_producto, $post->valor_producto);
+                if ($nombre == null) {
                     echo json_encode(
-                        array('message' => 'Error en ingreso para el inventario')
+                        array('message' => "no se encontro")
                     );
-                }else {
 
-
-                    $valor_total_inv =$di->valor_total(); //valor total de todo
-                    //multiplicar al cantidad por el valor
-                    //sacar el valor de ese producto
-                    $valor_del_prod = $post->valor_producto * $cantidad_total;
-                    $valor_total_inv +=  $valor_del_prod;
-                    if ($i-> actualizar_valor($valor_total_inv,$id_inventario)==false) {//me falta
+                    if ($di->create_detalle_inventario($post->nombre_producto, $cantidad_total, $post->valor_producto, 1, $p->id_bodega, $id_producto, $fecha) == false) { //listo
                         echo json_encode(
-                            array('message' => 'Error al actualizar el valor del inventario')
+                            array('message' => 'Error en ingreso para el inventario')
                         );
-                    }else {
+                    } else {
                         echo json_encode(
-                            array('message' => 'Post Created')
+                            array('message' => 'Se cre un nuevo producto')
                         );
                     }
+                } else {
+                    echo json_encode(
+                        array('message' => "Producto encontrado")
+                    );
+
+                    $di->id_detalle_inventario = $di->buscardor_igual_producto_id($post->nombre_producto, $post->valor_producto);
+                    $cantidad_d_i = $di->buscardor_igual_producto_cantidad($post->nombre_producto, $post->valor_producto);
+
+                    $cantidad_d_i = $cantidad_d_i + $cantidad_total;
+                    if ($di->Sumar_mismo_producto($di->id_detalle_inventario, $cantidad_d_i, $fecha) == false) {
+                        echo json_encode(
+                            array('message' => 'no se pudo actualizar el detalle del inventario')
+                        );
+                    } else {
+                        echo json_encode(
+                            array('message' => 'Se creo el actualizo el inventario del producto')
+                        );
+                    }
+                }
+
+
+
+
+
+
+
+                $valor_total = $di->valor_total();
+                $cantidades_total = $di->cantidad_total();
+                $total_inventario = $valor_total * $cantidades_total;
+
+
+                if ($i->actualizar_valor($valor_total, $total_inventario) == false) { //me falta
+                    echo json_encode(
+                        array('message' => 'Error al actualizar el valor del inventario')
+                    );
+                } else {
+                    echo json_encode(
+                        array('message' => 'Post Created')
+                    );
                 }
             }
         } else {
@@ -111,8 +143,6 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             );
         }
     }
-    
-    
 }
 
 if ($_SERVER['REQUEST_METHOD'] == 'GET') {
@@ -134,7 +164,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'GET') {
                     'producto.id_producto' => $post->id_producto,
                     'valor_producto' => $post->valor_producto,
                     'nombre_producto' => $post->nombre_producto,
-                    'cantidad_total' =>$post->cantidad_total
+                    'cantidad_total' => $post->cantidad_total
                 );
 
                 //Make JSON
@@ -183,30 +213,54 @@ if ($_SERVER['REQUEST_METHOD'] == 'DELETE') {
     // Instiate blog post object
     $post = new Controller_Producto($GLOBALS['db']);
     $po = new Controller_bodega_has_producto($GLOBALS['db']);
+    $di = new Controller_detalle_inventario($GLOBALS['db']);
 
     // GET ID
     $post->id_producto = isset($_GET['id_producto']) ? $_GET['id_producto'] : die();
+
+    $fecha = date('Y-m-d');
+
 
     if (!empty($post->buscar_id_producto($post->id_producto))) {
         echo json_encode(
             array('message' => 'no se encontro la bodega para eliminar')
         );
     } else {
-        if ($po->delete_bodega_has_producto($post->id_producto)) {
-            if ($post->delete_single_producto()) {
-                echo json_encode(
-                    array('message' => 'Post deleted')
-                );
-            }else {
+        //obtener cantidad del producto para despues restarla en el inventario
+        //cantidad total de ese producto en cuestion
+        $po->cantidad_total = $po->buscar_cantidad_producto($post->id_producto);
+        //obtenr nombre y valor  por id 
+        $post->nombre_producto = $post->Obtener_nombre_producto($post->id_producto);
+        $post->valor_producto = $post->Obtener_valor_producto($post->id_producto);
+
+        //restarla del inventario
+        //se obtien el total del inventario sobre el producto 
+        $cantidad_d_i = $di->buscardor_igual_producto_cantidad($post->nombre_producto, $post->valor_producto);
+        $cantidad_d_i = $cantidad_d_i - $po->cantidad_total;
+        //se obtiene el id
+        $di->id_detalle_inventario = $di->buscardor_igual_producto_id($post->nombre_producto, $post->valor_producto);
+
+
+        if ($di->Sumar_mismo_producto($di->id_detalle_inventario, $cantidad_d_i, $fecha) == false) {
+            echo json_encode(
+                array('message' => 'no se pudo actualizar el detalle del inventario')
+            );
+        } else {
+            if ($po->delete_bodega_has_producto($post->id_producto)) {
+                if ($post->delete_single_producto()) {
+                    echo json_encode(
+                        array('message' => 'Post deleted')
+                    );
+                } else {
+                    echo json_encode(
+                        array('message' => 'Post not deleted')
+                    );
+                }
+            } else {
                 echo json_encode(
                     array('message' => 'Post not deleted')
                 );
             }
-            
-        } else {
-            echo json_encode(
-                array('message' => 'Post not deleted')
-            );
         }
     }
 }
@@ -223,9 +277,10 @@ if ($_SERVER['REQUEST_METHOD'] == 'PUT') {
     $post->nombre_producto = $GLOBALS['data']->nombre_producto;
     $cantidad_total = $GLOBALS['data']->cantidad_total;
     $p->id_bodega = $GLOBALS['data']->id_bodega;
+    $fecha = date('Y-m-d');
     $validador = true;
-    
-    if ($post->buscar_id_producto($post->id_producto)==true) {
+
+    if ($post->buscar_id_producto($post->id_producto) == true) {
         $validador = false;
         echo json_encode(
             array('message' => 'No existe la bodega')
@@ -237,30 +292,44 @@ if ($_SERVER['REQUEST_METHOD'] == 'PUT') {
             array('message' => 'Existe mombre del Producto')
         );
     }
-    if ($post->Validador_valor_producto($post->valor_producto)!=null) {
-        $validador=false;
+    if ($post->Validador_valor_producto($post->valor_producto) != null) {
+        $validador = false;
         echo json_encode(
             array('message' => $post->Validador_valor_producto($post->valor_producto))
         );
     }
-    if ($po->Validador_cantidad_total($cantidad_total)!=null) {
-        $validador=false;
+    if ($po->Validador_cantidad_total($cantidad_total) != null) {
+        $validador = false;
         echo json_encode(
             array('message' => $po->Validador_cantidad_total($cantidad_total))
         );
     }
     if ($validador == true) {
         if ($post->update_producto()) {
-            if ($po->update_bodega_has_producto($p->id_bodega,$post->id_producto,$cantidad_total)) {
+            if ($po->update_bodega_has_producto($p->id_bodega, $post->id_producto, $cantidad_total)) {
+
+
+
+
+                $cantidad_d_i = $di->buscardor_igual_producto_cantidad($post->nombre_producto, $post->valor_producto);
+                $cantidad_d_i = $cantidad_d_i - $po->cantidad_total;
+                //se obtiene el id
+                $di->id_detalle_inventario = $di->buscardor_igual_producto_id($post->nombre_producto, $post->valor_producto);
+
+
+                if ($di->Sumar_mismo_producto($di->id_detalle_inventario, $cantidad_d_i, $fecha) == false) {
+                    echo json_encode(
+                        array('message' => 'no se pudo actualizar el detalle del inventario')
+                    );
+                }
                 echo json_encode(
                     array('message' => 'Post Update')
                 );
-            }else {
+            } else {
                 echo json_encode(
                     array('message' => 'Post not Update')
                 );
             }
-            
         } else {
             echo json_encode(
                 array('message' => 'Post not Update')
